requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.3.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
colorthief==0.2.1
pillow==10.1.0
python-dotenv==1.0.0
httpx==0.25.2
lxml==4.9.3
